requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
